+++ test/fixtures/parser/teradata/comment_on_column_stmt.sql
@@ -1,0 +1,2 @@
+comment on column sandbox_db.Org_Descendant.Org_Unit_Code is 'Organisational unit code';
+